firemark_django/requirements.txt
Django==3.1.7
django-filter>=2.4.0
djangorestframework==3.12.2
jsonfield==1.0.3
Markdown==2.6.2
six==1.9.0
wheel==0.24.0
json-spec
